More on the website

+++ WebservicesParent/LibrarySampler/WithJackrabbitOak/src/test/java/com/maiereni/oak/documentStore/derby/OakCleaner.java
@@ -15,43 +15,51 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.maiereni.oak.filestore;
+package com.maiereni.oak.documentStore.derby;
+
 
 import java.io.File;
+import java.sql.DriverManager;
 
 import org.apache.commons.io.FileUtils;
-import org.apache.jackrabbit.oak.segment.file.FileStore;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 import com.maiereni.oak.Cleaner;
-
+import com.maiereni.oak.bo.RepositoryProperties;
 /**
  * @author Petre Maierean
  *
  */
 @Component
-public class FolderCleaner implements Cleaner {
-	public static final Logger logger = LoggerFactory.getLogger(FolderCleaner.class);
-	
+public class OakCleaner implements Cleaner {
+	public static final Logger logger = LoggerFactory.getLogger(OakCleaner.class);
+
 	@Autowired
-	private File repositoryDir;
-	@Autowired
-	private FileStore segmentStore;
+	private RepositoryProperties repositoryProperties;
 	
-	@Override
 	public void cleanup() {
-		logger.debug("Start cleanup");
-		segmentStore.close();
-		try {
-			FileUtils.deleteDirectory(repositoryDir);
-			logger.debug("Directory has been deleted: " + repositoryDir.getPath());
-		}
-		catch(Exception e) {
-			logger.error("Failed to delete " + repositoryDir.getPath(), e);
+		logger.debug("Start cleaning up");
+		Object o = repositoryProperties.get(DerbyDocumentNodeStoreOakBeanFactory.DERBY_SYSTEM_HOME);
+		if (o != null) {
+			try {
+				String cleanup = (String)repositoryProperties.get(DerbyDocumentNodeStoreOakBeanFactory.DERBY_SHUTDOWN_KEY);
+				logger.debug("Cleanup " + cleanup);
+				DriverManager.getConnection(cleanup, "", "");
+			}
+			catch(Exception e) {
+				//logger.error("Cannot cleanup the datatabse", e);
+			}
+			File repositoryDir = (File)o;
+			try {
+				FileUtils.deleteDirectory(repositoryDir);
+				logger.debug("Directory has been deleted: " + repositoryDir.getPath());
+			}
+			catch(Exception e) {
+				logger.error("Failed to remove repository home: " + repositoryDir.getPath(), e);
+			}
 		}
 	}
-
 }
